Create 'ShowMoreAndBackToTopButtons' component
